Move _load_library to tensorflow_io.core.python.ops
_load_library used to be defined in tensorflow_io/__init__.py
which is not really desired as we actually want to use
tensorflow_io/__init__.py to expose "top level APIs"
