+++ tensorflow_io/__init__.py
@@ -17,43 +17,4 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-import os
-import sys
-import inspect
-
 import tensorflow as tf
-from tensorflow import errors
-
-def _load_library(filename, lib="op"):
-  """_load_library"""
-  f = inspect.getfile(sys._getframe(1)) # pylint: disable=protected-access
-
-  # Construct filename
-  f = os.path.join(os.path.dirname(f), filename)
-  filenames = [f]
-
-  # Add datapath to load if en var is set, used for running tests where shared
-  # libraries are built in a different path
-  datapath = os.environ.get('TFIO_DATAPATH')
-  if datapath is not None:
-    # Build filename from `datapath` + `package_name` + `relpath_to_library`
-    f = os.path.join(
-        datapath, __name__, os.path.relpath(f, os.path.dirname(__file__)))
-    filenames.append(f)
-
-  # Function to load the library, return True if file system library is loaded
-  load_fn = tf.load_op_library if lib == "op" \
-      else lambda f: tf.compat.v1.load_file_system_library(f) is None
-
-  # Try to load all paths for file, fail if none succeed
-  errs = []
-  for f in filenames:
-    try:
-      l = load_fn(f)
-      if l is not None:
-        return l
-    except errors.NotFoundError as e:
-      errs.append(str(e))
-  raise NotImplementedError(
-      "unable to open file: " +
-      "{}, from paths: {}\ncaused by: {}".format(filename, filenames, errs))
