+++ tensorflow_io/core/python/ops/__init__.py
@@ -17,5 +17,48 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-from tensorflow_io import _load_library
+import os
+import sys
+import inspect
+
+import tensorflow as tf
+
+def _load_library(filename, lib="op"):
+  """_load_library"""
+  f = inspect.getfile(sys._getframe(1)) # pylint: disable=protected-access
+
+  # Construct filename
+  f = os.path.join(os.path.dirname(f), filename)
+  filenames = [f]
+
+  # Add datapath to load if en var is set, used for running tests where shared
+  # libraries are built in a different path
+  datapath = os.environ.get('TFIO_DATAPATH')
+  if datapath is not None:
+    # Build filename from:
+    # `datapath` + `tensorflow_io` + `package_name` + `relpath_to_library`
+    rootpath = os.path.dirname(sys.modules['tensorflow_io'].__file__)
+    filename = sys.modules[__name__].__file__
+    f = os.path.join(
+        datapath, "tensorflow_io",
+        os.path.relpath(os.path.dirname(filename), rootpath),
+        os.path.relpath(f, os.path.dirname(filename)))
+    filenames.append(f)
+  # Function to load the library, return True if file system library is loaded
+  load_fn = tf.load_op_library if lib == "op" \
+      else lambda f: tf.compat.v1.load_file_system_library(f) is None
+
+  # Try to load all paths for file, fail if none succeed
+  errs = []
+  for f in filenames:
+    try:
+      l = load_fn(f)
+      if l is not None:
+        return l
+    except tf.errors.NotFoundError as e:
+      errs.append(str(e))
+  raise NotImplementedError(
+      "unable to open file: " +
+      "{}, from paths: {}\ncaused by: {}".format(filename, filenames, errs))
+
 core_ops = _load_library('libtensorflow_io.so')
