+++ tensorflow_io/gcs/python/ops/gcs_config_ops.py
@@ -26,7 +26,7 @@ from tensorflow.python.framework import dtypes
 from tensorflow.python.framework import ops
 from tensorflow.python.ops import array_ops
 from tensorflow.python.training import training
-from tensorflow_io import _load_library
+from tensorflow_io.core.python.ops import _load_library
 
 # Some GCS operations may be pre-defined and available via tf.contrib in
 # earlier TF versions. Because these ops are pre-registered, they will not be
