+++ tensorflow_io/grpc/python/ops/grpc_ops.py
@@ -19,7 +19,7 @@ from __future__ import print_function
 
 import tensorflow as tf
 from tensorflow.compat.v1 import data
-from tensorflow_io import _load_library
+from tensorflow_io.core.python.ops import _load_library
 grpc_ops = _load_library('_grpc_ops.so')
 
 class GRPCDataset(data.Dataset):
