+++ tensorflow_io/kafka/python/ops/kafka_dataset_ops.py
@@ -20,7 +20,7 @@ from __future__ import print_function
 import tensorflow as tf
 from tensorflow import dtypes
 from tensorflow.compat.v1 import data
-from tensorflow_io import _load_library
+from tensorflow_io.core.python.ops import _load_library
 kafka_ops = _load_library('_kafka_ops.so')
 
 class KafkaDataset(data.Dataset):
