+++ tensorflow_io/kafka/python/ops/kafka_ops.py
@@ -17,7 +17,7 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-from tensorflow_io import _load_library
+from tensorflow_io.core.python.ops import _load_library
 kafka_ops = _load_library('_kafka_ops.so')
