+++ tensorflow_io/oss/python/ops/ossfs_ops.py
@@ -22,6 +22,6 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-from tensorflow_io import _load_library
+from tensorflow_io.core.python.ops import _load_library
 
 _load_library("_oss_ops.so")
